// packages/asset-swapper/src/utils/market_operation_utils/types.ts
export enum ERC20BridgeSource {
    Native = 'Native',
    Uniswap = 'Uniswap',
    UniswapV2 = 'Uniswap_V2',
    Eth2Dai = 'Eth2Dai',
    Kyber = 'Kyber',
    Curve = 'Curve',
    LiquidityProvider = 'LiquidityProvider',
    MultiBridge = 'MultiBridge',
    Balancer = 'Balancer',
}

// tslint:disable: enum-naming
/**
 * Curve contract function selectors.
 */
export enum CurveFunctionSelectors {
    None = '0x00000000',
    exchange = '0x3df02124',
    exchange_underlying = '0xa6417ed6',
    get_dy_underlying = '0x07211ef7',
    get_dx_underlying = '0x0e71d1b9',
    get_dy = '0x5e0d443f',
    get_dx = '0x67df02ca',
}
// tslint:enable: enum-naming

/**
 * Configuration info on a Curve pool.
 */
export interface CurveInfo {
    exchangeFunctionSelector: CurveFunctionSelectors;
    sellQuoteFunctionSelector: CurveFunctionSelectors;
    buyQuoteFunctionSelector: CurveFunctionSelectors;
    poolAddress: string;
    tokens: string[];
}

// Internal `fillData` field for `Fill` objects.
export interface FillData {}

// `FillData` for native fills.
export interface NativeFillData extends FillData {
    order: SignedOrderWithFillableAmounts;
}

export interface CurveFillData extends FillData {
    fromTokenIdx: number;
    toTokenIdx: number;
    curve: CurveInfo;
}

export interface BalancerFillData extends FillData {
    poolAddress: string;
}

export interface UniswapV2FillData extends FillData {
    tokenAddressPath: string[];
}

export interface LiquidityProviderFillData extends FillData {
    poolAddress: string;
}

export interface MultiBridgeFillData extends FillData {
    poolAddress: string;
}

/**
 * Represents an individual DEX sample from the sampler contract.
 */
export interface DexSample<TFillData extends FillData = FillData> {
    source: ERC20BridgeSource;
    input: BigNumber;
    output: BigNumber;
    fillData?: TFillData;
}
// ...
export interface BatchedOperation<TResult> {
    encodeCall(): string;
    handleCallResultsAsync(callResults: string): Promise<TResult>;
}

export interface SourceQuoteOperation<TFillData extends FillData = FillData> extends BatchedOperation<BigNumber[]> {
    readonly source: ERC20BridgeSource;
    fillData?: TFillData;
}

export class SamplerContractOperation<TParams extends any[], TReturn, TFillData extends FillData = FillData>
    implements SourceQuoteOperation<TFillData> {
    constructor(
        private readonly _samplerContract: ERC20BridgeSamplerContract,
        public readonly source: ERC20BridgeSource,
        private readonly _samplerFunction: (...params: TParams) => ContractFunctionObj<TReturn>,
        private readonly _params: any[],
        public fillData: TFillData = {} as TFillData, // tslint:disable-line:no-object-literal-type-assertion
        public handleCallResultsAsync: (callResults: string) => Promise<BigNumber[]> = async (callResults: string) => {
            return _samplerContract.getABIDecodedReturnData<BigNumber[]>(_samplerFunction.name, callResults);
        },
    ) {}

    public encodeCall(): string {
        return this._samplerFunction
            .bind(this._samplerContract)(...(this._params as TParams))
            .getABIEncodedTransactionData();
    }
}